+++ src/lib/index.js
@@ -1,8 +1,3 @@
-import AgGrid, {
-    getApi,
-    getColumnApi,
-    getApiAsync,
-    getColumnApiAsync,
-} from './components/AgGrid.react';
+import AgGrid, {getApi, getApiAsync} from './components/AgGrid.react';
 
-export {AgGrid, getApi, getColumnApi, getApiAsync, getColumnApiAsync};
+export {AgGrid, getApi, getApiAsync};