+++ CHANGELOG.md
@@ -6,9 +6,24 @@ Links "DE#nnn" prior to version 2.0 point to the Dash Enterprise closed-source D
 
 ## UNRELEASED
 
-### Changed 
+### Added
+- [#246](https://github.com/plotly/dash-ag-grid/pull/246/)
+  - `grid_version` added to allow developer to see underlying AG Grid version in python
+  - Added `quartz` theme native support
+  - Added function support for `dateParser`, `dateFormatter`, `quickFilterParser`, `components`, `quickFilterMatcher`, `predicate`, `textFormatter`, `textMatcher`, `numberFormatter`, `numberParser`, `dataTypeMatcher`
 
+### Removed
+- [#246](https://github.com/plotly/dash-ag-grid/pull/246/) dropped `getColumnApi` and `getColmunApiAsync` as these are deprecated from the underlying grid.
+
+### Changed
 - [#261](https://github.com/plotly/dash-ag-grid/pull/261) The `cellValueChanged` property has changed been changed from a (single) event object to a _list_ of event objects. For multi-cell edits, the list will contain an element per change. In other cases, the list will contain a single element. Fixes [#262](https://github.com/plotly/dash-ag-grid/issues/262) 
+- [#246](https://github.com/plotly/dash-ag-grid/pull/246/)
+  - updating underlying grid version from AG Grid v29 -> v31, with this change, Dash AG Grid will reflect a similar version number to underlying Grid version.
+  - `dataTypeDefinitions` now supports full js, partial python with full definitions in js of an object, and objects have parts that are js
+  - `columnSize` and `columnState` cannot be currently passed together when grid initializes, the `columnSize` will trump the `columnState`
+
+### Fixed
+- [#246](https://github.com/plotly/dash-ag-grid/pull/246/) testing for grid going to destroyed state
 
 ## [2.4.0] - 2023-10-17